requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
Flask==1.1.1
Flask-Limiter==1.2.1
Flask-Login==0.5.0
idna==2.9
imutils==0.5.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
limits==1.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
numpy==1.18.2
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
pylint==2.4.4
requests==2.23.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
